# feat(query): optimize count(no_null_col)