I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

If count a not null column, I think we can optimize it as count(no_args).

From this case, we can see: the speed of count(no_null_col) is the same as count(), and count(nullable_col) is not affected.

```sql
mysql> desc orders;
+-----------------+---------+------+---------+-------+
| Field           | Type    | Null | Default | Extra |
+-----------------+---------+------+---------+-------+
| o_orderkey      | BIGINT  | NO   | 0       |       |
| o_custkey       | BIGINT  | NO   | 0       |       |
| o_orderstatus   | VARCHAR | NO   |         |       |
| o_totalprice    | DOUBLE  | NO   | 0       |       |
| o_orderdate     | DATE    | NO   | 0       |       |
| o_orderpriority | VARCHAR | NO   |         |       |
| o_clerk         | VARCHAR | NO   |         |       |
| o_shippriority  | INT     | NO   | 0       |       |
| o_comment       | VARCHAR | NO   |         |       |
+-----------------+---------+------+---------+-------+
9 rows in set (0.00 sec)
Read 0 rows, 0.00 B in 0.001 sec., 0 rows/sec., 0.00 B/sec.

mysql> desc orders_null;
+-----------------+---------+------+---------+-------+
| Field           | Type    | Null | Default | Extra |
+-----------------+---------+------+---------+-------+
| o_orderkey      | BIGINT  | YES  | NULL    |       |
| o_custkey       | BIGINT  | YES  | NULL    |       |
| o_orderstatus   | VARCHAR | YES  | NULL    |       |
| o_totalprice    | DOUBLE  | YES  | NULL    |       |
| o_orderdate     | DATE    | YES  | NULL    |       |
| o_orderpriority | VARCHAR | YES  | NULL    |       |
| o_clerk         | VARCHAR | YES  | NULL    |       |
| o_shippriority  | INT     | YES  | NULL    |       |
| o_comment       | VARCHAR | YES  | NULL    |       |
+-----------------+---------+------+---------+-------+
9 rows in set (0.01 sec)
Read 0 rows, 0.00 B in 0.001 sec., 0 rows/sec., 0.00 B/sec.


mysql> select count() from orders_null;
+----------+
| count()  |
+----------+
| 15000000 |
+----------+
1 row in set (0.01 sec)
Read 1 rows, 1.00 B in 0.001 sec., 842.2 rows/sec., 842.20 B/sec.

mysql> select count(o_comment) from orders_null;
+------------------+
| count(o_comment) |
+------------------+
|         15000000 |
+------------------+
1 row in set (0.27 sec)
Read 15000000 rows, 809.92 MiB in 0.261 sec., 57.47 million rows/sec., 3.03 GiB/sec.

mysql> select count() from orders;
+----------+
| count()  |
+----------+
| 15000000 |
+----------+
1 row in set (0.01 sec)
Read 1 rows, 1.00 B in 0.002 sec., 532.36 rows/sec., 532.36 B/sec.

mysql> select count(o_comment) from orders;
+------------------+
| count(o_comment) |
+------------------+
|         15000000 |
+------------------+
1 row in set (0.01 sec)
Read 1 rows, 1.00 B in 0.001 sec., 1.4 thousand rows/sec., 1.36 KiB/sec.

```

Closes #8614